sixtracklib_base/cobjects: adds complete C header only implementation for handling flat cobjects buffers
NOTE: flat buffers are linear chunks of memory that contain elements stored via the cobjects
buffer mechanism but with the memory slap not being managed / enclosed in any supporting structure
like a CBufferView or CBuffer.

This used to be called a "managed buffer" in earlier versions of sixtracklib, but this caused
confusion (the other buffers can also be managed). Flat buffers should be a better name